pyproject.toml
[project]
name = "character-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.133.1",
    "pycharacterai>=2.2.50",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
